requirements.txt
# Core AI and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.21.0
pillow>=9.0.0

# Web Framework
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# Data Processing and Visualization
matplotlib>=3.5.0
scipy>=1.9.0
pandas>=1.5.0
seaborn>=0.11.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
requests>=2.28.0

# YOLOv5 Dependencies
ultralytics>=8.0.0

# Optional: GPU Support (uncomment if you have CUDA)
# torch>=2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision>=0.15.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html